feat(model): add Jupyter Notebook with *keras* model
Add a test model that takes data and labels from *data_15000.npy* and output's prediction. The architecture is inspired from a [paper](https://github.com/twerkmeister/iLID/blob/master/Deep%20Audio%20Paper%20Thomas%20Werkmeister%2C%20Tom%20Herold.pdf). The model was only trained for a few seconds.
In front of the model, *kapre* is used to preprocess the data to Melspectrogramms at runtime.
The model was saved for later experimental use with the webserver.

IMPORTANT RESOURCES:

- https://github.com/twerkmeister/iLID/blob/master/Deep%20Audio%20Paper%20Thomas%20Werkmeister%2C%20Tom%20Herold.pdf
- https://github.com/keunwoochoi/kapre
- *DEEP LEARNING with Python* from *François Chollet* (book)